[ENH] [FIX] Moved the class FourierComponents into a separate components.py module in the certification. Fixed 'tabulate' for certification.
